pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nfl-betting-advisor"
version = "0.1.0"
description = "Command-line advisor for evaluating NFL parlay bets"
authors = [{name = "Generated by ChatGPT"}]
requires-python = ">=3.10"
dependencies = [
    "google-generativeai",
    "rich",
    "python-dotenv",
]
readme = "README.md"
keywords = ["nfl", "betting", "sports", "parlay"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Games/Entertainment :: Sports",
]

[project.scripts]
nfl-betting-advisor = "nfl_betting_advisor.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
